Upgrade log4j to version 2.15.0
Due to security vulnerability found in older versions of log4j, it is
required to upgrade it to 2.15.0. This dependency is directly added to the
pom file due to spring boot dependency management which brings the old
version regardless of what version we use in structured-logging. This
dependency would need to be removed once we upgrade spring boot to the
required patch 2.5.8 or 2.6.2. More information about this can be found
here https://spring.io/blog/2021/12/10/log4j2-vulnerability-and-spring-boot